requirements.txt

torch
torchvision
timm

pandas
scikit-learn  
joblib        

fastapi
pydantic      
uvicorn

langchain-google-genai
langchain-core
langchain-experimental
langgraph
langchain
python-dotenv

tabulate
python-multipart

ultralytics

